Redirect to user page with clear text explaining viewer has no permission
